project

+++ app/controllers/ProjectController.java
@@ -1,6 +1,5 @@
 package controllers;
 
-import com.alibaba.fastjson.JSON;
 import com.google.gson.Gson;
 import models.Condition;
 import models.Project;
@@ -16,28 +15,62 @@ import java.util.Map;
  */
 public class ProjectController extends Controller {
 
+    /**
+     * 状态码列表
+     *
+     * @param projectId
+     */
     public static void condition(String projectId) {
-        List<Condition> conditionList = Condition.getConditionListByProject(projectId);
-        render(conditionList, projectId);
+        if (StringUtils.isNotEmpty(projectId)) {
+            List<Condition> conditionList = Condition.getConditionListByProject(projectId);
+            Project project = Project.findById(projectId);
+            render(conditionList, projectId, project);
+        } else {
+            redirect("/404");
+        }
     }
 
+    /**
+     * 状态码修改、保存
+     *
+     * @param code
+     * @param describe
+     * @param projectId
+     * @param id
+     */
     public static void codeSave(String code, String describe, String projectId, String id) {
         if (StringUtils.isNotEmpty(id)) {
             Condition condition = Condition.findById(id);
             condition.code = code;
             condition.describe = describe;
             condition.save();
-        } else if (StringUtils.isNotEmpty(projectId)) {
-            Condition.saveCondition(code, describe, projectId);
+        } else {
+            if (StringUtils.isNotEmpty(projectId) && StringUtils.isNotEmpty(code)) {
+                Condition.saveCondition(code, describe, projectId);
+            }
         }
         condition(projectId);
     }
 
-    public static void delete(String id,String projectId) {
-        Condition.deleteById(id);
+    /**
+     * 删除状态码
+     *
+     * @param id
+     * @param projectId
+     */
+    public static void delete(String id, String projectId) {
+        if (StringUtils.isNotEmpty(id)) {
+            Condition.deleteById(id);
+        }
         condition(projectId);
     }
 
+    /**
+     * 根据id获取状态码
+     *
+     * @param id
+     * @return
+     */
     public static String getCondition(String id) {
         if (StringUtils.isNotEmpty(id)) {
             return getJsonString((Condition) Condition.findById(id));
@@ -45,13 +78,25 @@ public class ProjectController extends Controller {
         return "";
     }
 
+    /**
+     * 项目列表
+     */
     public static void project() {
+        //todo 根据用户获取自己的项目
         List<Project> projectList = Project.getProjeList();
         render(projectList);
     }
 
+    /**
+     * 保存项目
+     *
+     * @param name
+     * @param describe
+     */
     public static void save(String name, String describe) {
-        Project.projectSave(name, describe);
+        if (StringUtils.isNotEmpty(name)) {
+            Project.projectSave(name, describe);
+        }
         project();
     }
